deploy gitlab pages with the right baseURL
Hugo *requires* some baseURL setting. By default it's `/` which breaks
the site when deployed in a subpath, like when we deploy in gitlab
pages.

So, instead of reusing the CI template, we just vendor the code in
here and build hugo for pages specially. We can still have a more
"standard" staging as that will work because it's published at the
root, but at least pages will work now.

See: tpo/web/dev#15